Add insertOrUpdateAll method for bulk upsert support.
This differs from iteratively calling `insertOrUpdate` because it takes
place in a single JDBC batch.

NB: This is affected by https://github.com/pgjdbc/pgjdbc/issues/783
If the JDBC connection is configured with reWriteBatchedInserts=true,
then the postgres driver will mangle the generated SQL into being
invalid. There is no workaround (other than setting 'false'), but
there's an upstream PR which fixes the bug, so the window for that to
affect things is small, and it only affects people who opt into the new
API this exposes.

I haven't been able to figure out a reasonable fallback yet for when the
capability is disabled; using DBIO.sequence(insertOrUpdate).transactionally is
reasonable for semantics but doesn't compile due to DBIOAction != ProfileAction